requirements.txt
aiofiles~=23.1.0
aiohttp~=3.8.4
python-dotenv~=1.0.0
tiktoken~=0.4.0
